Move commonality to top level class
