+++ starter-cli/src/main/java/io/micronaut/starter/cli/command/BuilderCommand.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017-2022 original authors
+ * Copyright 2017-2023 original authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,15 +19,20 @@ import io.micronaut.core.util.StringUtils;
 import io.micronaut.core.util.SupplierUtil;
 import io.micronaut.starter.application.ApplicationType;
 import io.micronaut.starter.application.Project;
+import io.micronaut.starter.application.generator.ProjectGenerator;
 import io.micronaut.starter.feature.AvailableFeatures;
 import io.micronaut.starter.feature.BaseAvailableFeatures;
 import io.micronaut.starter.feature.Feature;
+import io.micronaut.starter.io.FileSystemOutputHandler;
+import io.micronaut.starter.io.OutputHandler;
 import io.micronaut.starter.options.BuildTool;
 import io.micronaut.starter.options.JdkVersion;
 import io.micronaut.starter.options.Language;
 import io.micronaut.starter.options.MicronautJdkVersionConfiguration;
+import io.micronaut.starter.options.Options;
 import io.micronaut.starter.options.TestFramework;
 import io.micronaut.starter.util.NameUtils;
+import org.fusesource.jansi.AnsiConsole;
 import org.jline.reader.EndOfFileException;
 import org.jline.reader.LineReader;
 import org.jline.reader.LineReaderBuilder;
@@ -35,10 +40,12 @@ import org.jline.reader.UserInterruptException;
 import org.jline.reader.impl.DefaultParser;
 import org.jline.reader.impl.completer.StringsCompleter;
 import org.jline.terminal.Terminal;
+import org.jline.terminal.TerminalBuilder;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Set;
 import java.util.concurrent.Callable;
 import java.util.function.Function;
 import java.util.function.Supplier;
@@ -52,6 +59,48 @@ import static picocli.CommandLine.Help.Ansi.AUTO;
  * @since 3.8.0
  */
 public abstract class BuilderCommand extends BaseCommand implements Callable<Integer> {
+
+    protected final ProjectGenerator projectGenerator;
+    protected final List<Feature> features;
+
+    protected BuilderCommand(ProjectGenerator projectGenerator, List<Feature> features) {
+        this.projectGenerator = projectGenerator;
+        this.features = features;
+    }
+
+    protected abstract GenerateOptions createGenerateOptions(LineReader reader);
+
+    @Override
+    public Integer call() throws Exception {
+        AnsiConsole.systemInstall();
+        try {
+            Terminal terminal = TerminalBuilder.terminal();
+            LineReader reader = LineReaderBuilder.builder()
+                    .terminal(terminal)
+                    .parser(new DefaultParser())
+                    .build();
+            GenerateOptions options = createGenerateOptions(reader);
+            Set<String> selectedFeatures = options.getFeatures();
+            selectedFeatures.addAll(getFeatures(options.getApplicationType(), terminal, features));
+            Project project = getProject(reader);
+            try (OutputHandler outputHandler = new FileSystemOutputHandler(project, false, this)) {
+                projectGenerator.generate(options.getApplicationType(),
+                        project,
+                        options.getOptions(),
+                        getOperatingSystem(),
+                        new ArrayList<>(selectedFeatures),
+                        outputHandler,
+                        this);
+                out("@|blue ||@ Application created at " + outputHandler.getOutputLocation());
+            }
+        } catch (UserInterruptException | EndOfFileException e) {
+            //no-op
+        } finally {
+            AnsiConsole.systemUninstall();
+        }
+        return 0;
+    }
+
     // Wrapped in a memoized Supplier, so it's not created at build time
     public static final Supplier<String> PROMPT = SupplierUtil.memoized(() -> AUTO.string("@|blue > |@"));
 
@@ -203,6 +252,23 @@ public abstract class BuilderCommand extends BaseCommand implements Callable<Int
                 err(e.getMessage());
             }
         }
+    }
+
+    protected Options getOptions(LineReader reader) {
+        Language language = getLanguage(reader);
+        TestFramework testFramework = getTestFramework(reader, language);
+        BuildTool buildTool = getBuildTool(reader, language);
+        JdkVersion jdkVersion = getJdkVersion(reader);
+        return new Options(language, testFramework, buildTool, jdkVersion);
+    }
+
+    protected Language getLanguage(LineReader reader) {
+        out("Choose your preferred language. (enter for default)");
+        return getEnumOption(
+                Language.class,
+                language -> StringUtils.capitalize(language.getName()),
+                Language.DEFAULT_OPTION,
+                reader);
     }
 
     protected TestFramework getTestFramework(LineReader reader, Language language) {
