+++ starter-cli/src/main/java/io/micronaut/starter/cli/command/YesOrNo.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017-2022 original authors
+ * Copyright 2017-2023 original authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -22,9 +22,4 @@ package io.micronaut.starter.cli.command;
 public enum YesOrNo {
     YES,
     NO;
-
-    @Override
-    public String toString() {
-        return this.name();
-    }
 }
